refactor: unify all tiers to shared high-perf cluster config
All tier configs (lite through ultra) now share the same
runtime_high_perf cluster settings (clients_max=256,
journal_slot_count=1024, message_size_max=10MiB, block_size=1MiB),
differing only in capacity quotas (storage/RAM limits).

Key changes:
- config.zig: Refactor tiers to use with_capacity() over shared base
- Add configs.config_name for build-tier identification in tests
- Widen client_count from u8 to u16 across test infrastructure
  (cluster.zig, packet_simulator.zig, vopr.zig, 4 test files)
- Cap C client test buffers to 64KB to avoid stack overflow with
  10MiB message_size_max
- Update skip guards from journal_slot_count check to config_name
  check for lite-skipped cluster tests
- Update golden checksum in replica_format.zig
- Update trace/checkpoint snapshot files
- Fix test-concurrent-clients.sh port detection (use pre-generated
  ports instead of log scraping)
- Fix benchmark_lsm.sh bash 4+ compat (auto-fallback to Homebrew
  bash on macOS)
- Add tier-profiles.md, capacity limit tests, Go loadgen tool
- Update docs and build system for new tier model

All 1705 unit tests pass, all SDKs pass, E2E 3-node cluster pass,
chaos tests pass, 100/100 concurrent clients pass.

+++ src/testing/cluster.zig
@@ -81,7 +81,7 @@ pub fn ClusterType(comptime StateMachineType: anytype) type {
             cluster_id: u128,
             replica_count: u8,
             standby_count: u8,
-            client_count: u8,
+            client_count: u16,
             storage_size_limit: u64,
             reformats_max: u8,
             seed: u64,
@@ -245,7 +245,7 @@ pub fn ClusterType(comptime StateMachineType: anytype) type {
             // There may be more clients than `clients_max` (to test session eviction).
             // +1 is for pulse which uses client_id = 0.
             const pipeline_requests_limit =
-                (@min(options.cluster.client_count, constants.clients_max) + @as(u8, 1)) -|
+                (@min(options.cluster.client_count, constants.clients_max) + @as(u16, 1)) -|
                 constants.pipeline_prepare_queue_max;
 
             for (replica_pools, 0..) |*pool, i| {
